refactor(cli): extract session command subsystem module (#164)
